requirements.txt
moviepy==1.0.3
pandas
wikipedia
requests
instagrapi
python-dotenv